rework MuiIcons
Icons don't have children
put icons in separate files
