+++ client/src/main/scala/com/example/pages/duplicate/DuplicatePageBridgeAppBar.scala
@@ -219,7 +219,7 @@ object DuplicatePageBridgeAppBarInternal {
 //                          MuiIcons.Check(
 //                              color= (if (HomePage.isFastclickOn) SvgColor.inherit else SvgColor.disabled),
 //                              classes = js.Dictionary("root" -> "mainMenuItemIcon").asInstanceOf[js.Object]
-//                          )()
+//                          )
 //                      ),
                       BeepComponent.getMenuItem( ()=>scope.withEffectsImpure.forceUpdate ),
                   )
