--- materialui/src/main/scala/com/example/materialui/icons/MuiIcons.scala
@@ -1,58 +1,0 @@
-package com.example.materialui.icons
-
-import scala.scalajs.js
-import scala.scalajs.js.annotation.JSImport
-import japgolly.scalajs.react.JsComponent
-import japgolly.scalajs.react.Children
-
-object MuiIcons {
-
-//  object Check extends SvgIconBase {
-//    @js.native @JSImport("@material-ui/icons/Check", JSImport.Default)
-//    private object icon extends js.Any
-//    protected val f = JsComponent[SvgIconProps, Children.Varargs, Null](icon)
-//  }
-
-  object CheckBox extends SvgIconBase {
-    @js.native @JSImport("@material-ui/icons/CheckBox", JSImport.Default)
-    private object icon extends js.Any
-    protected val f = JsComponent[SvgIconProps, Children.Varargs, Null](icon)
-  }
-
-  object CheckBoxOutlineBlank extends SvgIconBase {
-    @js.native @JSImport("@material-ui/icons/CheckBoxOutlineBlank", JSImport.Default)
-    private object icon extends js.Any
-    protected val f = JsComponent[SvgIconProps, Children.Varargs, Null](icon)
-  }
-
-  object ChevronRight extends SvgIconBase {
-    @js.native @JSImport("@material-ui/icons/ChevronRight", JSImport.Default)
-    private object icon extends js.Any
-    protected val f = JsComponent[SvgIconProps, Children.Varargs, Null](icon)
-  }
-
-  object Help extends SvgIconBase {
-    @js.native @JSImport("@material-ui/icons/Help", JSImport.Default)
-    private object icon extends js.Any
-    protected val f = JsComponent[SvgIconProps, Children.Varargs, Null](icon)
-  }
-
-  object Home extends SvgIconBase {
-    @js.native @JSImport("@material-ui/icons/Home", JSImport.Default)
-    private object icon extends js.Any
-    protected val f = JsComponent[SvgIconProps, Children.Varargs, Null](icon)
-  }
-
-  object Menu extends SvgIconBase {
-    @js.native @JSImport("@material-ui/icons/Menu", JSImport.Default)
-    private object icon extends js.Any
-    protected val f = JsComponent[SvgIconProps, Children.Varargs, Null](icon)
-  }
-
-  object MoreVert extends SvgIconBase {
-    @js.native @JSImport("@material-ui/icons/MoreVert", JSImport.Default)
-    private object icon extends js.Any
-    protected val f = JsComponent[SvgIconProps, Children.Varargs, Null](icon)
-  }
-
-}
